working on refactoring scrapers/employees
