+++ package.json
@@ -80,11 +80,6 @@
     "@babel/runtime": "^7.0.0",
     "@elastic/elasticsearch": "^7.0.0-rc.2",
     "@prisma/client": "^2.27.0",
-    "@types/cheerio": "^0.22.30",
-    "@types/history": "^4.7.6",
-    "@types/jsonwebtoken": "^8.5.4",
-    "@types/react-router": "^5.1.8",
-    "@types/request": "^2.48.7",
     "amplitude": "^4.0.1",
     "apollo-server": "^2.9.5",
     "apollo-server-express": "^2.9.7",
@@ -190,12 +185,23 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.8.3",
     "@types/atob": "^2.1.2",
+    "@types/cheerio": "^0.22.30",
+    "@types/cookie": "^0.4.1",
+    "@types/domutils": "^1.7.4",
+    "@types/fs-extra": "^9.0.13",
+    "@types/history": "^4.7.9",
+    "@types/htmlparser2": "^3.10.3",
     "@types/jest": "^25.1.4",
+    "@types/jsonwebtoken": "^8.5.6",
     "@types/lodash": "^4.14.150",
+    "@types/msgpack5": "^3.4.2",
     "@types/node": "^13.9.0",
     "@types/react": "^16.9.23",
     "@types/react-dom": "^16.9.5",
+    "@types/react-router": "^5.1.17",
     "@types/react-router-dom": "^5.1.3",
+    "@types/request": "^2.48.7",
+    "@types/urijs": "^1.19.17",
     "@types/webpack": "^4.41.8",
     "@typescript-eslint/eslint-plugin": "^2.23.0",
     "@typescript-eslint/parser": "^2.23.0",
