+++ types/scraperClasses.ts
